requirements.txt
pandas
numpy
matplotlib
librosa
seaborn
sklearn
IPython
tkinter
pillow
pyinstaller